+++ e-shop-app/src/App.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import './App.css';
-import Header from './app/shared/components/layout/Header'
-import Footer from './app/shared/components/layout/Footer'
-import Home from './app/pages/home/Home'
+import Header from './app/shared/components/layout/Header/Header'
+import Footer from './app/shared/components/layout/Footer/Footer'
+import Home from './app/pages/home'
 
 function App() {
   return (
