+++ e-shop-app/src/app/assets/images/index.tsx
