--- e-shop-app/src/app/pages/home/PolicyList.tsx
@@ -1,10 +1,0 @@
-import Policy from "./Policy";
-
-const PolicyList = () => {
- return(
-  <ul className="row policy-list">
-    <Policy/>
-  </ul>
- )
-}
-export default PolicyList;
