+++ e-shop-app/src/app/pages/home/components/Banner/index.tsx
@@ -1,0 +1,13 @@
+import campaignData from "../../../../shared/constants/CampaignData";
+import Banner from "./Banner";
+
+const Index = () => {
+  return(
+    <section className="section section-banner">
+      <>
+        <Banner campaign={campaignData}/>
+      </>
+    </section>
+  )
+}
+export default Index;
