+++ e-shop-app/src/app/pages/home/components/Blogs/index.tsx
@@ -1,0 +1,10 @@
+import BlogList from "./BlogList";
+
+const Index = () => {
+  return (
+    <section className="section section-blogs">
+      <BlogList />
+    </section>
+  )
+}
+export default Index;
