+++ e-shop-app/src/app/pages/home/components/Form/index.tsx
@@ -1,0 +1,12 @@
+import Form from "./Form";
+
+const Index = () => {
+  return (
+    <section className="section section-banner-form">
+      <div className="container ">
+        <Form />
+      </div>
+    </section>
+  )
+}
+export default Index;
