--- e-shop-app/src/app/shared/components/layout/Footer.tsx
@@ -1,106 +1,0 @@
-const Footer = () => {
-  return (
-    <footer className="section-footer text-secondary">
-      <div className="container footer">
-        <div className="footer-top">
-          <div className="footer-left">
-            <div className="footer-description">
-              <a className="footer-link" href="#">
-                <svg width={138} height={35} viewBox="0 0 138 35" fill="none" xmlns="http://www.w3.org/2000/svg">
-                  <path d="M74.8596 11.7466H67.6233V14.6504H74.2659V17.0254H67.6233V20.5415H75.1936V23H64.8308V9.3252H74.8596V11.7466ZM76.4089 16.4502H81.7341V18.9551H76.4089V16.4502ZM85.8162 18.7788C85.9028 19.4035 86.0728 19.8704 86.3264 20.1797C86.7903 20.7425 87.585 21.0239 88.7107 21.0239C89.3848 21.0239 89.9322 20.9497 90.3528 20.8013C91.1506 20.5168 91.5496 19.988 91.5496 19.2148C91.5496 18.7633 91.3516 18.4139 90.9558 18.1665C90.56 17.9253 89.9384 17.7119 89.0911 17.5264L87.6438 17.2017C86.2213 16.88 85.2379 16.5306 84.6936 16.1533C83.7721 15.5225 83.3113 14.536 83.3113 13.1938C83.3113 11.9692 83.7566 10.9518 84.6472 10.1416C85.5378 9.33138 86.8459 8.92627 88.5715 8.92627C90.0126 8.92627 91.2403 9.30973 92.2546 10.0767C93.2751 10.8374 93.8101 11.9445 93.8596 13.3979H91.1135C91.064 12.5754 90.7053 11.9909 90.0374 11.6445C89.592 11.4157 89.0385 11.3013 88.3767 11.3013C87.6407 11.3013 87.0531 11.4497 86.614 11.7466C86.1749 12.0435 85.9553 12.4578 85.9553 12.9897C85.9553 13.4784 86.1718 13.8433 86.6047 14.0845C86.8831 14.2453 87.4768 14.4339 88.386 14.6504L90.7424 15.2163C91.7753 15.4637 92.5546 15.7946 93.0803 16.209C93.8967 16.8522 94.3049 17.783 94.3049 19.0015C94.3049 20.2508 93.8256 21.2899 92.8669 22.1187C91.9145 22.9412 90.5662 23.3525 88.822 23.3525C87.0408 23.3525 85.6399 22.9474 84.6194 22.1372C83.5989 21.3208 83.0886 20.2013 83.0886 18.7788H85.8162ZM101.782 12.6279C102.407 12.6279 102.973 12.7362 103.48 12.9526C103.994 13.1691 104.414 13.5 104.742 13.9453C105.02 14.3226 105.19 14.7122 105.252 15.1143C105.314 15.5101 105.345 16.1595 105.345 17.0625V23H102.645V16.8491C102.645 16.3049 102.552 15.8657 102.367 15.5317C102.126 15.0617 101.668 14.8267 100.994 14.8267C100.295 14.8267 99.7631 15.0617 99.3982 15.5317C99.0395 15.9956 98.8601 16.6605 98.8601 17.5264V23H96.2253V9.37158H98.8601V14.2051C99.2436 13.6175 99.6858 13.2093 100.187 12.9805C100.694 12.7454 101.226 12.6279 101.782 12.6279ZM112.359 21.1353C113.132 21.1353 113.726 20.86 114.14 20.3096C114.554 19.7591 114.761 18.9767 114.761 17.9624C114.761 16.9481 114.554 16.1688 114.14 15.6245C113.726 15.0741 113.132 14.7988 112.359 14.7988C111.586 14.7988 110.989 15.0741 110.568 15.6245C110.154 16.1688 109.947 16.9481 109.947 17.9624C109.947 18.9767 110.154 19.7591 110.568 20.3096C110.989 20.86 111.586 21.1353 112.359 21.1353ZM117.535 17.9624C117.535 19.4468 117.109 20.7178 116.255 21.7754C115.402 22.8268 114.106 23.3525 112.368 23.3525C110.63 23.3525 109.334 22.8268 108.481 21.7754C107.627 20.7178 107.2 19.4468 107.2 17.9624C107.2 16.5028 107.627 15.238 108.481 14.168C109.334 13.098 110.63 12.563 112.368 12.563C114.106 12.563 115.402 13.098 116.255 14.168C117.109 15.238 117.535 16.5028 117.535 17.9624ZM126.405 17.9438C126.405 17.1646 126.225 16.4749 125.866 15.875C125.514 15.2751 124.939 14.9751 124.141 14.9751C123.182 14.9751 122.524 15.4297 122.165 16.3389C121.979 16.8213 121.886 17.4336 121.886 18.1758C121.886 19.3509 122.199 20.1766 122.823 20.6528C123.195 20.9312 123.634 21.0703 124.141 21.0703C124.877 21.0703 125.437 20.7858 125.82 20.2168C126.21 19.6478 126.405 18.8901 126.405 17.9438ZM124.902 12.665C126.083 12.665 127.082 13.1011 127.898 13.9731C128.721 14.8452 129.132 16.1255 129.132 17.814C129.132 19.5952 128.73 20.9528 127.926 21.8867C127.128 22.8206 126.098 23.2876 124.837 23.2876C124.033 23.2876 123.365 23.0866 122.833 22.6846C122.542 22.4619 122.258 22.1372 121.979 21.7104V26.98H119.363V12.8877H121.896V14.3813C122.18 13.9422 122.483 13.5959 122.805 13.3423C123.392 12.8908 124.091 12.665 124.902 12.665Z" fill="white" />
-                  <path d="M19.4211 35C19.1691 35.009 18.9208 34.9303 18.711 34.7749C18.5013 34.6194 18.3405 34.395 18.2511 34.133L15.3171 25.577H2.421C2.10292 25.577 1.78796 25.5074 1.49411 25.3721C1.20027 25.2368 0.933288 25.0385 0.708454 24.7885C0.48362 24.5385 0.305335 24.2417 0.183774 23.9151C0.0622137 23.5885 -0.000235823 23.2385 6.69174e-07 22.885V2.692C-0.000119585 2.33777 0.0627081 1.98703 0.184873 1.65984C0.307038 1.33265 0.48615 1.03546 0.711915 0.785355C0.937679 0.535248 1.20565 0.337114 1.50049 0.202347C1.79532 0.0675792 2.11121 -0.00116969 2.43001 1.50554e-05H19.3887C19.6186 4.3819e-06 19.8438 0.0726692 20.0379 0.209549C20.232 0.346429 20.387 0.541881 20.4849 0.773026L24.1479 9.42303H38.7747C39.0928 9.42303 39.4077 9.49267 39.7016 9.62798C39.9954 9.76329 40.2624 9.9616 40.4873 10.2116C40.7121 10.4616 40.8904 10.7584 41.0119 11.085C41.1335 11.4116 41.1959 11.7616 41.1957 12.115V32.307C41.1958 32.6613 41.133 33.012 41.0108 33.3392C40.8887 33.6664 40.7096 33.9636 40.4838 34.2137C40.258 34.4638 39.9901 34.6619 39.6952 34.7967C39.4004 34.9315 39.0845 35.0002 38.7657 34.999L19.4211 35ZM18.6156 2.692H2.43001C2.43001 2.692 2.43001 9.42302 2.43001 22.884H27.1746L18.6156 2.692ZM25.2846 12.115L30.1734 23.657C30.2854 23.9216 30.3168 24.2194 30.2629 24.5056C30.2089 24.7917 30.0727 25.0506 29.8746 25.243L22.6089 32.307H38.7657C38.7657 32.307 38.7657 25.576 38.7657 12.115H25.2846ZM19.8972 31.366L25.8525 25.576H17.91L19.8972 31.366Z" fill="white" />
-                </svg>
-              </a>
-              <p className="link-content">House My Brand designs clothing for the young, the old &amp;everyone in between – but
-                most importantly, for the fashionable</p>
-            </div>
-            <ul className="social-list">
-              <li className="social-item">
-                <a className="social-link" href="#">
-                  <i className="fa fa-facebook" />
-                </a>
-              </li>
-              <li className="social-item">
-                <a className="social-link" href="#">
-                  <i className="fa fa-twitter" />
-                </a>
-              </li>
-              <li className="social-item">
-                <a className="social-link" href="#">
-                  <i className="fa fa-linkedin" />
-                </a>
-              </li>
-              <li className="social-item">
-                <a className="social-link" href="#">
-                  <i className="fa fa-instagram" />
-                </a>
-              </li>
-              <li className="social-item">
-                <a className="social-link" href="#">
-                  <i className="fa fa-youtube-play" />
-                </a>
-              </li>
-            </ul>
-          </div>
-          <ul className="footer-right">
-            <li className="footer-item">
-              <h5 className="item-title">Shopping online</h5>
-              <ul className="item-list">
-                <li className="sub-item">
-                  <a href="#" className="sub-item-link">Order Status</a>
-                </li>
-                <li className="sub-item">
-                  <a href="#" className="sub-item-link">Shipping and Delivery</a>
-                </li>
-                <li className="sub-item">
-                  <a href="#" className="sub-item-link">Returns</a>
-                </li>
-                <li className="sub-item">
-                  <a href="#" className="sub-item-link">Payment Options</a>
-                </li>
-                <li className="sub-item">
-                  <a href="#" className="sub-item-link">Contact Us</a>
-                </li>
-              </ul>
-            </li>
-            <li className="footer-item">
-              <h5 className="item-title">Information</h5>
-              <ul className="item-list">
-                <li className="sub-item">
-                  <a href="#" className="sub-item-link">Gift Cards</a>
-                </li>
-                <li className="sub-item">
-                  <a href="#" className="sub-item-link">Find a store</a>
-                </li>
-                <li className="sub-item">
-                  <a href="#" className="sub-item-link">Newsletter</a>
-                </li>
-                <li className="sub-item">
-                  <a href="#" className="sub-item-link">Bacome a member</a>
-                </li>
-                <li className="sub-item">
-                  <a href="#" className="sub-item-link">Site feedback</a>
-                </li>
-              </ul>
-            </li>
-            <li className="footer-item">
-              <h5 className="item-title">Contact</h5>
-              <ul className="item-list">
-                <li className="sub-item">
-                  <a href="#" className="sub-item-link">store@uikit.com</a>
-                </li>
-                <li className="sub-item">
-                  <a href="#" className="sub-item-link">Hotline: +1 131 138 138</a>
-                </li>
-              </ul>
-            </li>
-          </ul>
-        </div>
-      </div>
-      <div className="footer-bottom">
-        <p className="copyright">DESIGN BY ICEO.CO - © 2019. ALL RIGHTS RESERVED.</p>
-      </div>
-    </footer>
-  )
-}
-export default Footer;
