+++ e-shop-app/src/app/shared/components/layout/Footer/AboutUs/AboutUs.tsx
@@ -1,0 +1,4 @@
+const AboutUs = () => {
+  
+}
+export default AboutUs;
