+++ e-shop-app/src/app/shared/components/layout/Footer/Social/Social.tsx
@@ -1,0 +1,33 @@
+const Social = () => {
+  return (
+    <>
+      <li className="social-item">
+        <a className="social-link" href="#">
+          <i className="fa fa-facebook" />
+        </a>
+      </li>
+      <li className="social-item">
+        <a className="social-link" href="#">
+          <i className="fa fa-twitter" />
+        </a>
+      </li>
+      <li className="social-item">
+        <a className="social-link" href="#">
+          <i className="fa fa-linkedin" />
+        </a>
+      </li>
+      <li className="social-item">
+        <a className="social-link" href="#">
+          <i className="fa fa-instagram" />
+        </a>
+      </li>
+      <li className="social-item">
+        <a className="social-link" href="#">
+          <i className="fa fa-youtube-play" />
+        </a>
+      </li>
+    </>
+  )
+}
+
+export default Social;
