+++ e-shop-app/src/app/shared/components/layout/Footer/Social/SocialList.tsx
@@ -1,0 +1,11 @@
+import Social from "./Social";
+
+const SocialList = () => {
+  return(
+    <ul className="social-list">
+      <Social/>
+    </ul>
+  )
+}
+
+export default SocialList;
