+++ e-shop-app/src/app/shared/components/layout/Header/CategoriesList.tsx
@@ -1,0 +1,11 @@
+import Category from "./Category";
+
+const CategoryList = () => {
+  return(
+    <ul className="menu sub-menu-list">
+      <Category/>
+    </ul>
+  )
+}
+
+export default CategoryList;
