+++ e-shop-app/src/app/shared/components/layout/Header/Category.tsx
@@ -1,0 +1,16 @@
+const Category = () => {
+  return (
+    <>
+      <li className="sub-menu-item">
+        <a className="sub-menu-link text-primary" href="#">Men</a>
+      </li>
+      <li className="sub-menu-item">
+        <a className="sub-menu-link text-primary" href="#">Women</a>
+      </li>
+      <li className="sub-menu-item">
+        <a className="sub-menu-link text-primary" href="#">Kids</a>
+      </li>
+    </>
+  )
+}
+export default Category;
