+++ e-shop-app/src/app/shared/constants/BlogData.tsx
@@ -1,0 +1,13 @@
+import Images from "../../assets/images";
+
+const BlogData = [
+  {
+    id: 1,
+    title: "Free Shipping",
+    description: "On purchases over $199"
+  },
+  
+
+]
+
+export default BlogData;
