+++ e-shop-app/src/app/shared/constants/CampaignData.tsx
@@ -1,0 +1,25 @@
+import Images from "../../assets/images";
+
+const CampaignData = [
+  {
+    id: 1,
+    title: "Free Shipping",
+    description: "On purchases over $199",
+    imgSrc: Images.Shipping1
+  },
+  {
+    id: 2,
+    title: "99% Satisfied Customers",
+    description: "Our clients' opinions speak for themselves",
+    imgSrc: Images.Happy
+  },
+  {
+    id: 3,
+    title: "Originality Guaranteed",
+    description: "30 days warranty for each product from our store",
+    imgSrc: Images.Guarantee1
+  },
+
+]
+
+export default CampaignData;
