+++ e-shop-app/src/app/shared/constants/CategoriesData.tsx
@@ -1,0 +1,4 @@
+const CategoriesData = () => {
+
+}
+export default CategoriesData;
