+++ e-shop-app/src/app/shared/constants/PolicyData.tsx
@@ -1,0 +1,29 @@
+import Images from "../../assets/images";
+
+export const PolicyData = [
+  {
+    id: 1,
+    title: "Free Shipping",
+    description: "All purchases over $199 are eligible for free shipping via USPS First Class Mail.",
+    imgSrc: Images.Shipping2
+  },
+  {
+    id: 2,
+    title: "Easy Payments",
+    description: "All payments are processed instantlyover a secure payment protocol.",
+    imgSrc: Images.Payment
+  },
+  {
+    id: 3,
+    title: "Money-Back Guarantee",
+    description: "If an item arrived damaged or you'vechanged your mind, you can send itback for a full refund.",
+    imgSrc: Images.Guarantee2
+  },
+  {
+    id: 4,
+    title: "Finest Quality",
+    description: "Designed to last, each of our products hasbeen crafted with the finest materials.",
+    imgSrc: Images.Materials
+  },
+  
+]
