+++ e-shop-app/src/app/shared/interfaces/product.tsx
@@ -1,0 +1,7 @@
+export interface IProduct {
+  id: string,
+  name: string,
+  price: number,
+  discount: number,
+  imgSrc: string,
+}
