+++ e-shop-app/src/stylesheet/base/_variables.scss
@@ -14,27 +14,27 @@ $color: (
 // Font-size
 
 $font-size: (
-  font-xxs-3: 8px,
-  font-xxs-2: 10px,
-  font-xxs: 12px,
-  font-xs: 13px,
-  font-sm: 14px,
-  font-md: 16px,
-  font-md-2: 18px,
-  font-md-3: 20px,
-  font-lg: 24px,
-  font-xl: 30px,
-  font-xl-2: 32px,
-  font-xl-3: 48px,
-  font-xxl: 72px,
+  xxs-1: 8px,
+  xxs-2: 10px,
+  xxs-3: 12px,
+  xs: 13px,
+  sm: 14px,
+  md-1: 16px,
+  md-2: 18px,
+  md-3: 20px,
+  lg: 24px,
+  xl-1: 30px,
+  xl-2: 32px,
+  xl-3: 48px,
+  xxl: 72px,
 );
 
 //Font weight
 
 $font-weight: (
-  font-w300: 300,
-  font-w400: 400,
-  font-w700: 700,
+  w300: 300,
+  w400: 400,
+  w700: 700,
 );
 
 //Font family
