+++ e-shop-app/src/stylesheet/layout/_grid.scss
@@ -1,3 +1,5 @@
+$grid-columns : 12;
+
 .container {
   margin: 0 auto;
   width: 100%;
@@ -15,44 +17,12 @@
   }
 }
 
-.col {
-  &-1 {
-    width: 8.33%;
-  }
-  &-2 {
-    width: 16.67%;
-  }
-  &-3 {
-    width: 25%;
-  }
-  &-4 {
-    width: 33.33%;
-  }
-  &-5 {
-    width: 41.67%;
-  }
-  &-6 {
-    width: 50%;
-  }
-  &-7 {
-    width: 58.33%;
-  }
-  &-8 {
-    width: 66.67%;
-  }
-  &-9 {
-    width: 75%;
-  }
-  &-10 {
-    width: 83.33%;
-  }
-  &-11 {
-    width: 91.67%;
-  }
-  &-12 {
-    width: 100%;
+@for $i from 1 through $grid-columns {
+  .col-#{$i} {
+    max-width: percentage($i / $grid-columns);
   }
 }
+
 
 @include respond-to-mobile{
   .container {
@@ -69,10 +39,10 @@
     }
   }
 
-  .col-sm-12 {
-    width: 100%;
-  }
-  .col-sm-6 {
-    width: 50%;
+  @for $i from 1 through $grid-columns {
+    .col-sm-#{$i} {
+      max-width: percentage($i / $grid-columns);
+    }
   }
 }
+
