+++ e-shop-app/src/stylesheet/layout/_header.scss
@@ -20,8 +20,8 @@
 
   &.sub-menu-list {
     gap: 0 96px;
-    font-size: map-get($font-size, font-md-2);
-    font-weight: map-get($font-weight, font-w700);
+    font-size: map-get($font-size, md-2);
+    font-weight: map-get($font-weight, w700);
     line-height: 27px;
 
     &:hover {
