+++ e-shop-app/src/stylesheet/modules/_button.scss
@@ -3,7 +3,7 @@
   padding: 14px;
   border: none;
   border-radius: 33px;
-  font-size: map-get($font-size,font-xs);
+  font-size: map-get($font-size,xs);
   font-weight: bold;
   text-align: center;
   text-transform: uppercase;
